add a getNetworkInfo method that accepts an rpc and chain id and returns all the network info available for us
