+++ src/controllers/settings/settings.ts
@@ -1,13 +1,5 @@
-import { Contract, JsonRpcProvider } from 'ethers'
+import { JsonRpcProvider } from 'ethers'
 
-import EntryPointAbi from '../../../contracts/compiled/EntryPoint.json'
-import {
-  AMBIRE_ACCOUNT_FACTORY,
-  AMBIRE_PAYMASTER,
-  ERC_4337_ENTRYPOINT,
-  OPTIMISTIC_ORACLE,
-  SINGLETON
-} from '../../consts/deploy'
 import { networks } from '../../consts/networks'
 import { Key } from '../../interfaces/keystore'
 import { AvailableFeature, NetworkDescriptor, NetworkId } from '../../interfaces/networkDescriptor'
@@ -21,8 +13,8 @@ import {
 } from '../../interfaces/settings'
 import { Storage } from '../../interfaces/storage'
 import { getSASupport } from '../../libs/deployless/simulateDeployCall'
+import { getNetworkInfo } from '../../libs/settings/settings'
 import { isValidAddress } from '../../services/address'
-import { Bundler } from '../../services/bundlers/bundler'
 import EventEmitter from '../eventEmitter/eventEmitter'
 
 export class SettingsController extends EventEmitter {
@@ -313,46 +305,24 @@ export class SettingsController extends EventEmitter {
     }
 
     try {
-      const provider = new JsonRpcProvider(customNetwork.rpcUrl)
-      const [entryPointCode, singletonCode, oracleCode, factoryCode, hasBundler, block, saSupport] =
-        await Promise.all([
-          provider.getCode(ERC_4337_ENTRYPOINT),
-          provider.getCode(SINGLETON),
-          provider.getCode(OPTIMISTIC_ORACLE),
-          provider.getCode(AMBIRE_ACCOUNT_FACTORY),
-          Bundler.isNetworkSupported(customNetwork.chainId),
-          provider.getBlock('latest'),
-          getSASupport(provider)
-        ])
-      const has4337 = entryPointCode !== '0x' && hasBundler
-      const areContractsDeployed = factoryCode !== '0x'
-      let hasPaymaster = false
-      if (has4337) {
-        const entryPoint = new Contract(ERC_4337_ENTRYPOINT, EntryPointAbi, provider)
-        const paymasterBalance = await entryPoint.balanceOf(AMBIRE_PAYMASTER)
-        hasPaymaster = paymasterBalance.toString() > 0
-      }
-      const erc4337 = { erc4337: { enabled: has4337, hasPaymaster } }
-      const feeOptions = { feeOptions: { is1559: block?.baseFeePerGas !== null } }
+      const {
+        isSAEnabled,
+        isOptimistic,
+        hasSimulations,
+        erc4337,
+        areContractsDeployed,
+        feeOptions
+      } = await getNetworkInfo(customNetwork.rpcUrl, customNetwork.chainId)
 
-      // Ambire support is as follows:
-      // - either the addresses match after simulation, that's perfect
-      // - or we can't do the simulation with this RPC but we have the factory
-      // deployed on the network
-      const supportsAmbire =
-        saSupport.addressMatches || (!saSupport.supportsStateOverride && areContractsDeployed)
-
-      const addCustomNetwork: CustomNetwork = {
+      this.#networkPreferences[customNetworkId] = {
         ...customNetwork,
         ...erc4337,
         ...feeOptions,
-        isSAEnabled: supportsAmbire && singletonCode !== '0x',
+        isSAEnabled,
         areContractsDeployed,
-        isOptimistic: oracleCode !== '0x',
-        hasSimulations: saSupport.supportsStateOverride
+        isOptimistic,
+        hasSimulations
       }
-
-      this.#networkPreferences[customNetworkId] = addCustomNetwork
     } catch (e: any) {
       this.emitError({
         message:
